Passing spec tests for partials
- Indentation and IdentationQueue

The mustache-js implementation has a very simple way to handle indentations:
it just renders the partial as string, splits by '\n' and append each part with the indentation.

Mustache-zig can't do that because there is no string buffer, everything is rendered directly into the writer.
So, we need a way to keep track of the last '\n' and nested indentations (there is no final indentation if the rendered  partial ends with a '\n').

- Iterator rework

Simpler way to handle iterators.
When next() is called it immediately tries to fetch the next element, making hasNext() very cheap to be called.

- Tests

Passing all specs tests
Added some extra tests (for indented partials, and partials inside iterators)