Fix segfaul expanding lambdas
The LambdaInvoker struct is intended to hold a pointer to the current context and the function's pointer to the lambda itself.
This struct is created on the stack and should be copied, instead it was passed by pointer, causing a segfault.

Still, it is not clear why it just crashed on Windows Operating System, not on Linux or MacOS.